Move function that may be used more generically to soss_utils.py
